restrict the view to be portrait

+++ CHZZSudoku/CHZZGridView.m
@@ -100,7 +100,7 @@
     [self resetColor];
     UIButton* button = (UIButton*) sender;
     [button setBackgroundColor:[UIColor colorWithWhite:1 alpha:0.5]];
-    int buttonTag = button.tag;
+    int buttonTag = (int) button.tag;
     int row = buttonTag / 10;
     int col = buttonTag % 10;
     [_target performSelector:_action withObject:[NSNumber numberWithInt:row] withObject:[NSNumber numberWithInt:col]];
